+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [8.0.2] - 2019-11-05
+
+### Added
+- Listing BIM360 users
+- Searching BIM360 users using filters
+
 ## [8.0.1] - 2019-11-04
 
 ### Added